refactor: introduce job execution use cases
